Move some type definitions into more manageable modules.
