+++ rustfmt.toml
@@ -1,4 +1,4 @@
-edition="2021"
+edition = "2021"
 combine_control_expr = false
 wrap_comments = true
 brace_style = "PreferSameLine"
@@ -7,8 +7,8 @@ fn_single_line = true
 format_strings = true
 format_macro_matchers = true
 format_macro_bodies = true
-imports_granularity="Crate"
-normalize_comments = false
+imports_granularity = "Crate"
+normalize_comments = true
 reorder_impl_items = true
 reorder_imports = true
 struct_field_align_threshold = 20
@@ -19,4 +19,3 @@ use_try_shorthand = true
 format_code_in_doc_comments = true
 overflow_delimited_expr = true
 normalize_doc_attributes = true
-use_small_heuristics = "Off"
